#178 [refactor]: move admin exception handlers into routers/admin (register_admin_exception_handlers)
